feat(ui): add new components for error messages, editor area, and sidebar functionality
