+++ ConfigForge/Models/ErrorMessage.swift
@@ -1,0 +1,14 @@
+//
+//  ErrorMessage.swift
+//  ConfigForge
+//
+//  Created by samzong
+//
+
+import Foundation
+
+// 添加一个可识别的错误消息结构体
+struct ErrorMessage: Identifiable {
+    let id = UUID()
+    let message: String
+}
