+++ ConfigForge/Views/Components/EmptyEditorViewModern.swift
@@ -1,0 +1,35 @@
+//
+//  EmptyEditorViewModern.swift
+//  ConfigForge
+//
+//  Created by samzong
+//
+
+import SwiftUI
+
+// 现代化的空编辑器视图
+struct EmptyEditorViewModern: View {
+    var body: some View {
+        VStack(spacing: 20) {
+            Image(systemName: "square.on.square.dashed")
+                .font(.system(size: 56))
+                .foregroundColor(.secondary)
+                .symbolRenderingMode(.hierarchical)
+                .padding(.bottom, 10)
+            
+            LocalizedText("editor.empty.title")
+                .font(.title3)
+                .foregroundColor(.primary)
+            
+            LocalizedText("editor.empty.description")
+                .font(.subheadline)
+                .foregroundColor(.secondary)
+                .multilineTextAlignment(.center)
+                .fixedSize(horizontal: false, vertical: true)
+                .frame(maxWidth: 300)
+        }
+        .padding()
+        .frame(maxWidth: .infinity, maxHeight: .infinity)
+        .background(Color(.windowBackgroundColor))
+    }
+}
