+++ ConfigForge/Views/Components/MessageBanner.swift
@@ -1,0 +1,60 @@
+//
+//  MessageBanner.swift
+//  ConfigForge
+//
+//  Created by samzong
+//
+
+import SwiftUI
+
+// 消息横幅
+struct MessageBanner: View {
+    let message: AppMessage
+    let onDismiss: () -> Void
+    
+    var body: some View {
+        HStack(spacing: 10) {
+            Image(systemName: iconName)
+                .font(.system(size: 14, weight: .semibold))
+                .foregroundColor(.white)
+            
+            Text(message.message)
+                .font(.system(size: 13, weight: .medium))
+                .foregroundColor(.white)
+            
+            if message.type != .success {
+                Button(action: {
+                    // 直接调用 dismiss 操作
+                    onDismiss()
+                }) {
+                    Image(systemName: "xmark")
+                        .font(.system(size: 10, weight: .bold))
+                        .foregroundColor(.white.opacity(0.8))
+                }
+                .buttonStyle(PlainButtonStyle())
+                .padding(.leading, 4)
+            }
+        }
+        .padding(.vertical, 8)
+        .padding(.horizontal, 12)
+        .background(Capsule().fill(bannerColor))
+        .shadow(color: Color.black.opacity(0.15), radius: 3, x: 0, y: 2)
+        .padding(.top, 12)
+    }
+    
+    private var iconName: String {
+        switch message.type {
+        case .success: return "checkmark.circle.fill"
+        case .error: return "exclamationmark.triangle.fill"
+        case .info: return "info.circle.fill"
+        }
+    }
+    
+    private var bannerColor: Color {
+        switch message.type {
+        case .success: return Color.green.opacity(0.9)
+        case .error: return Color.red.opacity(0.9)
+        case .info: return Color.blue.opacity(0.9)
+        }
+    }
+}
